# News-Analysis-Engine/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DetectionRequest {
  /** News article title */
  title: string;
  /** Full news article text content */
  content: string;
  /** Base64 encoded image (optional) */
  imageBase64?: string;
  /** URL of image to analyze (optional) */
  imageUrl?: string;
  /** Original URL source (optional) */
  sourceUrl?: string;
}

export interface ModelScore {
  /** Name of the model */
  modelName: string;
  /** Type/role of the model */
  modelType: string;
  /** Score from 0 to 1 (0=real, 1=fake) */
  score: number;
  /** Human-readable label */
  label: string;
  /** Brief explanation of the score */
  details: string;
}

export interface DetectionResult {
  /** Unique detection ID */
  id: string;
  title: string;
  sourceUrl?: string;
  textAigcScore: ModelScore;
  textIntentScore: ModelScore;
  imageAigcScore: ModelScore;
  /** Final fused score from linear layer (0=real, 1=fake) */
  finalScore: number;
  /** Final verdict label */
  finalLabel: string;
  /** LLM-generated analysis report */
  report: string;
  analyzedAt: string;
}

export interface CrawlRequest {
  /** URL of the news article to crawl */
  url: string;
}

export interface CrawlResult {
  url: string;
  title: string;
  content: string;
  /** List of image URLs found in the article */
  imageUrls: string[];
  publishedAt?: string;
  author?: string;
  siteName?: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}