# News-Analysis-Engine/lib/api-zod/src/generated/types/crawlRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface CrawlRequest {
  /** URL of the news article to crawl */
  url: string;
}